add plex button component
